refactor(logs): flatten lib/ directory and split utils.ts into focused modules
- Move all files from domains/logs/lib/ to domains/logs/ (convention:
  domain files go directly in the domain root, not in a nested lib/)
- Split the 355-line catch-all utils.ts into:
  - format.ts: formatting utilities (timestamps, tokens, cost, latency)
  - trace-event-processing.ts: trace event grouping, metrics, icons, status
  - usage-tab-state.ts: absorbed resolveRangeWindow/resolveUsageGranularity
    (only used by usage tab)
- Remove unused eslint-disable directive in usage-tab.tsx
- Update all imports across the logs domain
