+++ apps/web/src/domains/logs/trace-events-types.ts
