+++ apps/web/src/domains/logs/usage-tab-state.ts
@@ -1,11 +1,12 @@
-import { LLM_USAGE_DIMENSION_LABELS } from "@/domains/logs/lib/llm-dimension.js";
-import { UsageRequestError } from "@/domains/logs/lib/usage-api.js";
+import { LLM_USAGE_DIMENSION_LABELS } from "./llm-dimension.js";
+import { UsageRequestError } from "./usage-api.js";
 import type {
   UsageGranularity,
   UsageGroupBy,
   UsageSeriesBucket,
   UsageSeriesGroupBy,
-} from "@/domains/logs/lib/usage-types.js";
+  UsageTimeRange,
+} from "./usage-types.js";
 
 export const DEFAULT_USAGE_GROUP_BY: UsageGroupBy = "task";
 export const FALLBACK_USAGE_GROUP_BY: UsageGroupBy = "model";
@@ -108,4 +109,39 @@ function isHourlyBucket(
 
 function isHourlyDate(date: string) {
   return /^\d{4}-\d{2}-\d{2} ([01]\d|2[0-3]):00$/.test(date);
+}
+
+const RANGE_START_DAY_OFFSETS: Record<Exclude<UsageTimeRange, "all">, number> =
+  {
+    today: 0,
+    "7d": 6,
+    "30d": 29,
+    "90d": 89,
+  };
+
+export function resolveRangeWindow(
+  range: UsageTimeRange,
+  now: Date | number = Date.now(),
+): {
+  from: number;
+  to: number;
+} {
+  const to = typeof now === "number" ? now : now.getTime();
+  if (range === "all") {
+    return { from: 0, to };
+  }
+  const localNow = new Date(to);
+  const dayOffset = RANGE_START_DAY_OFFSETS[range];
+  const from = new Date(
+    localNow.getFullYear(),
+    localNow.getMonth(),
+    localNow.getDate() - dayOffset,
+  ).getTime();
+  return { from, to };
+}
+
+export function resolveUsageGranularity(
+  range: UsageTimeRange,
+): UsageGranularity {
+  return range === "today" ? "hourly" : "daily";
 }
